Move the event_utils to common
tendrl-bug-id: Tendrl/node-agent#714
